added directory structure to node app w/ READMEs, migrated routing, formatted files
