+++ README.md
@@ -1,1 +1,7 @@
 The backend that powers the HACS website!
+
+Our current API Routes:
+
+- /officers
+- /calendar
+- /siteContent
