+++ models/README.md
@@ -1,0 +1,1 @@
+Place for our DB models
