+++ scripts/README.md
@@ -1,0 +1,1 @@
+Place for useful scripts to help development
